+++ CHANGELOG.md
@@ -1,1 +1,6 @@
- 
+## Changelog
+### 🐞 Bug Fixes
+- Fixed synchronization bug between the client and the server
+- Fixed blur error in debug screen / and other
+- Fixed LDL4 support
+- Fixed an issue where lantern does not drop upon death